refactor feed scoring logic
