+++ services/rank_score_feeds/helper.py
@@ -4,8 +4,6 @@ from datetime import timedelta
 import json
 import os
 
-import numpy as np
-
 from lib.aws.athena import Athena
 from lib.aws.dynamodb import DynamoDB
 from lib.aws.glue import Glue
@@ -30,6 +28,7 @@ from services.rank_score_feeds.models import (
     CustomFeedPost,
     ScoredPostModel,
 )
+from services.rank_score_feeds.scoring import calculate_post_scores
 
 max_feed_length = 50
 default_lookback_days = 3
@@ -37,11 +36,6 @@ consolidated_enriched_posts_table_name = "consolidated_enriched_post_records"
 user_to_social_network_map_table_name = "user_social_networks"
 feeds_root_s3_key = "custom_feeds"
 dynamodb_table_name = "rank_score_feed_sessions"
-
-default_similarity_score = 0.75
-average_popular_post_like_count = 250
-coef_toxicity = 0.965
-coef_constructiveness = 1.02
 
 athena = Athena()
 s3 = S3()
@@ -167,54 +161,6 @@ def export_post_scores(post_uri_to_post_score_map: dict[str, str]):
     output_json_body_bytes = output_jsons.encode("utf-8")
     key = os.path.join("post_scores", f"post_scores_{current_datetime_str}.jsonl")  # noqa
     s3.write_to_s3(blob=output_json_body_bytes, key=key)
-
-
-def calculate_post_score(post: ConsolidatedEnrichedPostModel) -> float:
-    """Calculate a post's score.
-
-    Calculates what a score's post would be, depending on whether or not it's
-    used for the engagement or the treatment condition.
-    """
-    engagement_score = None
-    treatment_score = None
-    engagement_coef = 1.0
-    if (
-        post.sociopolitical_was_successfully_labeled
-        and post.is_sociopolitical
-        and post.perspective_was_successfully_labeled
-    ):
-        if post.prob_constructive is None:
-            # backwards-compatbility bug where prob_constructive wasn't
-            # set, but the endpoints for prob_reasoning and prob_constructive
-            # are actually the same.
-            post.prob_constructive = post.prob_reasoning
-        # if sociopolitical, uprank/downrank based on toxicity/constructiveness.
-        treatment_coef = (
-            post.prob_toxic * coef_toxicity
-            + post.prob_constructive * coef_constructiveness
-        ) / (post.prob_toxic + post.prob_constructive)
-    else:
-        # if it's not sociopolitical, use a coef of 1.
-        treatment_coef = 1.0
-    if post.like_count:
-        engagement_score = engagement_coef * np.log(post.like_count + 1)
-        treatment_score = treatment_coef * np.log(post.like_count + 1)
-    elif post.similarity_score:
-        expected_like_count = average_popular_post_like_count * post.similarity_score
-        engagement_score = engagement_coef * np.log(expected_like_count + 1)
-        treatment_score = treatment_coef * np.log(expected_like_count + 1)
-    else:
-        # if we didn't have a like count or similarity score, return a
-        # score where we assume both of those.
-        expected_like_count = average_popular_post_like_count * default_similarity_score
-        engagement_score = engagement_coef * np.log(expected_like_count + 1)
-        treatment_score = treatment_coef * np.log(expected_like_count + 1)
-    return {"engagement_score": engagement_score, "treatment_score": treatment_score}
-
-
-def calculate_post_scores(posts: list[ConsolidatedEnrichedPostModel]) -> list[dict]:  # noqa
-    """Calculate scores for a list of posts."""
-    return [calculate_post_score(post) for post in posts]
 
 
 def calculate_in_network_posts_for_user(
